requirements.txt
apache-airflow-providers-snowflake
snowflake-connector-python
apache-airflow
WTForms==2.3.3
python-dotenv
scikit-learn
matplotlib
streamlit
yfinance
notebook
fredapi
pandas
numpy